- fixing zaza_utils. - added description of Analyze fields. - added analyze in the cli
