+++ cou/zaza_utils/os_versions.py
@@ -472,19 +472,3 @@ class CompareOpenStack(BasicStringComparator):
     """
 
     _list = list(OPENSTACK_CODENAMES.values())
-
-
-def determine_next_openstack_release(release):
-    """Determine the next release after the one passed as a str.
-
-    The returned value is a tuple of the form: ('2020.1', 'ussuri')
-
-    :param release: the release to use as the base
-    :type release: str
-    :returns: the release tuple immediately after the current one.
-    :rtype: Tuple[str, str]
-    :raises: KeyError if the current release doesn't actually exist
-    """
-    old_index = list(OPENSTACK_CODENAMES.values()).index(release)
-    new_index = old_index + 1
-    return list(OPENSTACK_CODENAMES.items())[new_index]
