+++ cou/zaza_utils/upgrade_utils.py
@@ -127,3 +127,19 @@ def get_upgrade_groups(model_name=None):
     apps_in_model = get_upgrade_candidates(model_name=model_name, filters=filters)
 
     return _build_service_groups(apps_in_model)
+
+
+def determine_next_openstack_release(release):
+    """Determine the next release after the one passed as a str.
+
+    The returned value is a tuple of the form: ('2020.1', 'ussuri')
+
+    :param release: the release to use as the base
+    :type release: str
+    :returns: the release tuple immediately after the current one.
+    :rtype: Tuple[str, str]
+    :raises: KeyError if the current release doesn't actually exist
+    """
+    old_index = list(os_versions.OPENSTACK_CODENAMES.values()).index(release)
+    new_index = old_index + 1
+    return list(os_versions.OPENSTACK_CODENAMES.items())[new_index]
